requirements.txt
kafka-python==2.0.2
Faker==12.0.0
